# newgram-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Newgram API",
    "description": "API para Newgram, uma aplicação de rede social inspirada no Instagram.",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Jameson Henrique",
      "url": "https://github.com/JamesonHenrique",
      "email": "jamesonhenrique14@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Registrar novo usuário",
        "description": "Cria um novo usuário no sistema e retorna um token de autenticação",
        "operationId": "registrar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UsuarioCreateDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/TokenDTO" } }
            }
          },
          "400": { "description": "Dados inválidos para registro" },
          "409": { "description": "Email ou nome de usuário já cadastrado" }
        }
      }
    },
    "/auth/refresh-token": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Renovar token",
        "description": "Renova o token de acesso utilizando um refresh token válido",
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "description": "Refresh token para renovação"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token renovado com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/TokenDTO" } }
            }
          },
          "401": { "description": "Refresh token inválido ou expirado" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Autenticar usuário",
        "description": "Autentica o usuário com email e senha e retorna um token de acesso",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Autenticação realizada com sucesso",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/TokenDTO" } }
            }
          },
          "401": { "description": "Credenciais inválidas" },
          "404": { "description": "Usuário não encontrado" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UsuarioCreateDTO": {
        "type": "object",
        "description": "Dados do usuário a ser registrado",
        "properties": {
          "nome": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "senha": { "type": "string" },
          "bio": { "type": "string" },
          "confirmacaoSenha": { "type": "string" }
        }
      },
      "TokenDTO": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "refreshToken": { "type": "string" }
        }
      },
      "LoginDTO": {
        "type": "object",
        "description": "Credenciais de login do usuário",
        "properties": {
          "email": { "type": "string" },
          "senha": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Insira um token JWT válido aqui. Expira em 24h.",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshToken": {
        "type": "http",
        "description": "Insira um refresh token válido para obter novos tokens. Validade de 7 dias.",
        "name": "refreshToken",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}